students/k3340/laboratory_works/Zakoulov_Ilya/backend/src/city_orient_backend/urls.py
from django.urls import path, include
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path('api/quests/', include('quests.api.endpoints')),
    path('api/users/', include('users.api.endpoints')),
    path('api/auth', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh', TokenRefreshView.as_view(), name='token_refresh'),
]
