students/k3340/laboratory_works/Zakoulov_Ilya/backend/requirements/base.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
